Moving my home screen code into the HomeScreen folder and moving home screen unit tests into the testing folder
